create rich amount input view model
